Improved operator[], added entry of two or more unary operators(!!false)

+++ Lotus/src/utils/lotusError.cpp
@@ -13,7 +13,7 @@ void lotus::throwOverloadError(const String& overload, const String& type1, cons
     throw LotusException(type1 + STRING_LITERAL(": ") + STRING_LITERAL("No overload for") + overload + STRING_LITERAL(" with ") + type2);
 }
 
-void lotus::checkThrowIndexError(const Value& index, size_t size) {
+void lotus::checkThrowIndexError(const Value& index, int size) {
     if (index->asInt() < 0 || index->asInt() >= size) {
         throw LotusException(INT(index->asInt())->asString() + STRING_LITERAL(" is invalid index"));
     }
